refactor(admin): redesign admin UI with shadcn components and design system
- Install shadcn Card component; use it for stat cards on AdminDashboardPage
- AdminDashboardPage: replace raw divs with Card/CardHeader/CardTitle/CardContent,
  swap hardcoded hex colors for Tailwind design-token classes (text-content-*,
  text-accent, bg-surface-*, etc.)
- AdminLayout: remove all inline style={{}} attributes; use bg-surface-secondary,
  border-border, text-accent, bg-surface-tertiary Tailwind classes throughout
- AdminUsersPage: replace hardcoded hex with semantic classes; split 614-line file
  into BanUserDialog, DeleteUserDialog, SetRoleDialog, SessionsPanel,
  UserBadges, UserTableRows, useToasts, types — main file reduced to 211 lines
- Extract shared User and DialogKind types to types.ts
- Zero inline style={{}} remain (except status bar dynamic width); no hardcoded
  hex color values in component code; all styling via Tailwind design-token classes
