+++ apps/web/src/routes/admin/AdminLayout.tsx
@@ -12,18 +12,21 @@ export function AdminLayout() {
   return (
     <div className="min-h-screen bg-surface-primary flex">
       {/* Sidebar */}
-      <aside className="w-56 shrink-0 fixed top-0 left-0 h-full flex flex-col" style={{ background: "#18181B", borderRight: "1px solid #27272A" }}>
-        <div className="px-4 py-4 border-b" style={{ borderColor: "#27272A" }}>
-          <button onClick={() => navigate("/")} className="flex items-center gap-1.5 text-xs text-zinc-400 hover:text-zinc-200 transition-colors">
+      <aside className="w-56 shrink-0 fixed top-0 left-0 h-full flex flex-col bg-surface-secondary border-r border-border">
+        <div className="px-4 py-4 border-b border-border">
+          <button
+            onClick={() => navigate("/")}
+            className="flex items-center gap-1.5 text-xs text-content-tertiary hover:text-content-primary transition-colors"
+          >
             <ChevronLeft size={14} />
             Back to Board
           </button>
         </div>
 
-        <div className="px-4 py-4 border-b" style={{ borderColor: "#27272A" }}>
+        <div className="px-4 py-4 border-b border-border">
           <div className="flex items-center gap-2">
-            <BarChart3 size={16} className="text-[#22D3EE]" />
-            <span className="text-sm font-semibold text-zinc-100 tracking-tight">Admin</span>
+            <BarChart3 size={16} className="text-accent" />
+            <span className="text-sm font-semibold text-content-primary tracking-tight">Admin</span>
           </div>
         </div>
 
@@ -34,17 +37,11 @@ export function AdminLayout() {
               to={to}
               end={end}
               className={({ isActive }) =>
-                `flex items-center gap-2.5 px-3 py-2 rounded-md text-sm transition-colors relative ${
-                  isActive ? "text-[#22D3EE] bg-[#27272A]" : "text-zinc-400 hover:text-zinc-200 hover:bg-[#27272A]"
+                `flex items-center gap-2.5 px-3 py-2 rounded-md text-sm transition-colors border-l-2 ${
+                  isActive
+                    ? "text-accent bg-surface-tertiary border-l-accent"
+                    : "text-content-tertiary hover:text-content-primary hover:bg-surface-tertiary border-l-transparent"
                 }`
-              }
-              style={({ isActive }) =>
-                isActive
-                  ? {
-                      borderLeft: "2px solid #22D3EE",
-                      paddingLeft: "calc(0.75rem - 2px)",
-                    }
-                  : { borderLeft: "2px solid transparent", paddingLeft: "calc(0.75rem - 2px)" }
               }
             >
               <Icon size={15} />
