+++ apps/web/src/routes/admin/useToasts.tsx
@@ -1,0 +1,39 @@
+import { useCallback, useState } from "react";
+
+export interface Toast {
+  id: number;
+  type: "success" | "error";
+  message: string;
+}
+
+let toastSeq = 0;
+
+export function useToasts() {
+  const [toasts, setToasts] = useState<Toast[]>([]);
+
+  const push = useCallback((type: Toast["type"], message: string) => {
+    const id = ++toastSeq;
+    setToasts((prev) => [...prev, { id, type, message }]);
+    setTimeout(() => setToasts((prev) => prev.filter((t) => t.id !== id)), 3500);
+  }, []);
+
+  return { toasts, push };
+}
+
+export function ToastStack({ toasts }: { toasts: Toast[] }) {
+  if (toasts.length === 0) return null;
+  return (
+    <div className="fixed bottom-4 right-4 z-[100] flex flex-col gap-2">
+      {toasts.map((t) => (
+        <div
+          key={t.id}
+          className={`px-4 py-2.5 rounded-lg text-sm font-medium shadow-lg ${
+            t.type === "success" ? "bg-success text-zinc-900" : "bg-error text-white"
+          }`}
+        >
+          {t.message}
+        </div>
+      ))}
+    </div>
+  );
+}
